# mobile/openapi.json
{
 "openapi": "3.1.0",
 "info": {
  "title": "Triplet API",
  "description": "Trip planning aplication API",
  "version": "1.0.0"
 },
 "paths": {
...
  "/users/me": {
   "get": {
    "tags": [
     "User Data"
    ],
    "summary": "Get Me",
    "operationId": "get_me_users_me_get",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserResponse"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   },
   "delete": {
    "tags": [
     "User Data"
    ],
    "summary": "Delete User",
    "operationId": "delete_user_users_me_delete",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AccountDelete"
       }
      }
     },
     "required": true
    },
    "responses": {
     "204": {
      "description": "Successful Response"
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   },
   "patch": {
    "tags": [
     "User Data"
    ],
    "summary": "Update Profile",
    "operationId": "update_profile_users_me_patch",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/UserUpdate"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/users/me/password/code": {
   "post": {
    "tags": [
     "User Data"
    ],
    "summary": "Send Password Code",
    "description": "For accounts made with Google or Apple: email a code to set a first password with. A code\nto the account's inbox, not just a signed-in session, so someone using a phone left\nunlocked can't give themselves a password to the account.",
    "operationId": "send_password_code_users_me_password_code_post",
    "responses": {
     "202": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/MessageResponse"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/users/me/password": {
   "post": {
    "tags": [
     "User Data"
    ],
    "summary": "Set Password",
    "description": "Set a first password with the emailed code. Signing in with Google or Apple keeps working.",
    "operationId": "set_password_users_me_password_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PasswordSet"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/users/me/email/verify": {
   "post": {
    "tags": [
     "User Data"
    ],
    "summary": "Verify New Email",
    "description": "Enter the code sent to a new email address; the account switches to it.",
    "operationId": "verify_new_email_users_me_email_verify_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CodeRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/users/me/avatar": {
   "put": {
    "tags": [
     "User Data"
    ],
    "summary": "Upload Avatar",
    "operationId": "upload_avatar_users_me_avatar_put",
    "requestBody": {
     "content": {
      "multipart/form-data": {
       "schema": {
        "$ref": "#/components/schemas/Body_upload_avatar_users_me_avatar_put"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserResponse"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   },
   "delete": {
    "tags": [
     "User Data"
    ],
    "summary": "Delete Avatar",
    "operationId": "delete_avatar_users_me_avatar_delete",
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/UserResponse"
        }
       }
      }
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ]
   }
  },
  "/users/{user_id}/avatar": {
   "get": {
    "tags": [
     "User Data"
    ],
    "summary": "Get Avatar",
    "operationId": "get_avatar_users__user_id__avatar_get",
    "parameters": [
     {
      "name": "user_id",
      "in": "path",
      "required": true,
      "schema": {
       "type": "integer",
       "title": "User Id"
      }
     },
     {
      "name": "v",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "maxLength": 64,
       "default": "",
       "title": "V"
      }
     },
     {
      "name": "sig",
      "in": "query",
      "required": false,
      "schema": {
       "type": "string",
       "maxLength": 64,
       "default": "",
       "title": "Sig"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {}
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
  "/users/search": {
   "get": {
    "tags": [
     "User Data"
    ],
    "summary": "Search Users",
    "operationId": "search_users_users_search_get",
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "parameters": [
     {
      "name": "q",
      "in": "query",
      "required": true,
      "schema": {
       "type": "string",
       "minLength": 2,
       "maxLength": 255,
       "title": "Q"
      }
     },
     {
      "name": "limit",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "maximum": 100,
       "minimum": 1,
       "default": 50,
       "title": "Limit"
      }
     },
     {
      "name": "offset",
      "in": "query",
      "required": false,
      "schema": {
       "type": "integer",
       "minimum": 0,
       "default": 0,
       "title": "Offset"
      }
     }
    ],
    "responses": {
     "200": {
      "description": "Successful Response",
      "content": {
       "application/json": {
        "schema": {
         "type": "array",
         "items": {
          "$ref": "#/components/schemas/UserPublic"
         },
         "title": "Response Search Users Users Search Get"
        }
       }
      }
     },
     "422": {
      "description": "Validation Error",
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      }
     }
    }
   }
  },
...
 },
 "components": {
  "schemas": {
   "AccountDelete": {
    "properties": {
     "password": {
      "type": "string",
      "maxLength": 128,
      "title": "Password"
     }
    },
    "type": "object",
    "required": [
     "password"
    ],
    "title": "AccountDelete"
   },
...
   "Body_upload_avatar_users_me_avatar_put": {
    "properties": {
     "file": {
      "type": "string",
      "contentMediaType": "application/octet-stream",
      "title": "File"
     }
    },
    "type": "object",
    "required": [
     "file"
    ],
    "title": "Body_upload_avatar_users_me_avatar_put"
   },
...
   "CodeRequest": {
    "properties": {
     "code": {
      "type": "string",
      "maxLength": 12,
      "title": "Code"
     }
    },
    "type": "object",
    "required": [
     "code"
    ],
    "title": "CodeRequest"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "type": "array",
      "title": "Detail"
     }
    },
    "type": "object",
    "title": "HTTPValidationError"
   },
...
   "MessageResponse": {
    "properties": {
     "detail": {
      "type": "string",
      "title": "Detail"
     }
    },
    "type": "object",
    "required": [
     "detail"
    ],
    "title": "MessageResponse"
   },
...
   "PasswordSet": {
    "properties": {
     "code": {
      "type": "string",
      "maxLength": 12,
      "title": "Code"
     },
     "password": {
      "type": "string",
      "maxLength": 128,
      "title": "Password"
     }
    },
    "type": "object",
    "required": [
     "code",
     "password"
    ],
    "title": "PasswordSet"
   },
...
   "UserPublic": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "avatar_url": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Avatar Url"
     }
    },
    "type": "object",
    "required": [
     "id",
     "name"
    ],
    "title": "UserPublic"
   },
   "UserResponse": {
    "properties": {
     "id": {
      "type": "integer",
      "title": "Id"
     },
     "name": {
      "type": "string",
      "title": "Name"
     },
     "email": {
      "type": "string",
      "format": "email",
      "title": "Email"
     },
     "avatar_url": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Avatar Url"
     },
     "pending_email": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Pending Email"
     },
     "has_password": {
      "type": "boolean",
      "title": "Has Password",
      "default": true
     }
    },
    "type": "object",
    "required": [
     "id",
     "name",
     "email"
    ],
    "title": "UserResponse"
   },
   "UserUpdate": {
    "properties": {
     "name": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 255
       },
       {
        "type": "null"
       }
      ],
      "title": "Name"
     },
     "email": {
      "anyOf": [
       {
        "type": "string",
        "format": "email"
       },
       {
        "type": "null"
       }
      ],
      "title": "Email"
     },
     "password": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 128
       },
       {
        "type": "null"
       }
      ],
      "title": "Password"
     },
     "current_password": {
      "anyOf": [
       {
        "type": "string",
        "maxLength": 128
       },
       {
        "type": "null"
       }
      ],
      "title": "Current Password"
     }
    },
    "type": "object",
    "title": "UserUpdate"
   },
   "ValidationError": {
    "properties": {
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "type": "array",
      "title": "Location"
     },
     "msg": {
      "type": "string",
      "title": "Message"
     },
     "type": {
      "type": "string",
      "title": "Error Type"
     },
     "input": {
      "title": "Input"
     },
     "ctx": {
      "type": "object",
      "title": "Context"
     }
    },
    "type": "object",
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError"
   },
...
  },
  "securitySchemes": {
   "HTTPBearer": {
    "type": "http",
    "scheme": "bearer"
   }
  }
 }
}